[benc/a11y-metrics] Move parsers to their own directory
